refactor: Move `CompactionError` into error file
