+++ src/compaction/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod error;
 pub(crate) mod leveled;
 use std::{pin::Pin, sync::Arc};
 
@@ -6,10 +7,10 @@ use fusio_parquet::writer::AsyncWriter;
 use futures_util::StreamExt;
 use leveled::LeveledCompactor;
 use parquet::arrow::AsyncArrowWriter;
-use thiserror::Error;
 use tokio::sync::oneshot;
 
 use crate::{
+    compaction::error::CompactionError,
     fs::{generate_file_id, FileType},
     inmem::immutable::{ArrowArrays, Builder},
     record::{KeyRef, Record, Schema as RecordSchema},
@@ -156,29 +157,6 @@ where
         });
         Ok(())
     }
-}
-
-#[derive(Debug, Error)]
-pub enum CompactionError<R>
-where
-    R: Record,
-{
-    #[error("compaction io error: {0}")]
-    Io(#[from] std::io::Error),
-    #[error("compaction parquet error: {0}")]
-    Parquet(#[from] parquet::errors::ParquetError),
-    #[error("compaction fusio error: {0}")]
-    Fusio(#[from] fusio::Error),
-    #[error("compaction version error: {0}")]
-    Version(#[from] VersionError<R>),
-    #[error("compaction logger error: {0}")]
-    Logger(#[from] fusio_log::error::LogError),
-    #[error("compaction channel is closed")]
-    ChannelClose,
-    #[error("database error: {0}")]
-    Commit(#[from] CommitError<R>),
-    #[error("the level being compacted does not have a table")]
-    EmptyLevel,
 }
 
 #[cfg(all(test, feature = "tokio"))]
